Add clap-based CLI to exercise generator
