+++ rust-tooling/Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
-members = ["generator", "ci-tests", "utils", "models"]
+members = ["generate", "ci-tests", "utils", "models"]
 resolver = "2"
