+++ rust-tooling/generate/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "generator"
+name = "generate"
 version = "0.1.0"
 edition = "2021"
 description = "Generates exercise boilerplate, especially test cases"
@@ -7,6 +7,7 @@ description = "Generates exercise boilerplate, especially test cases"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+clap = { version = "4.4.8", features = ["derive"] }
 convert_case = "0.6.0"
 glob = "0.3.1"
 inquire = "0.6.2"
