+++ rust-tooling/generate/src/exercise_generation.rs
